src/third_party/WebKit/LayoutTests/gamepad/full-screen-gamepad-expected.txt

CONSOLE MESSAGE: line 37: This test requires User Activation V2.
END OF TEST
